+++ js/ui/camera.js
@@ -8,50 +8,49 @@ var LatLngBounds = require('../geo/lat_lng_bounds');
 var Point = require('point-geometry');
 
 /**
- * @typedef {Object} [animOptions]
- * @param {Number} [animOptions.duration=500] Number in milliseconds
- * @param {Function} animOptions.easing
- * @param {Array} [animOptions.offset=[0,0]] point, origin of movement relative to map center
- * @param {Boolean} [animOptions.animate=true] When set to false, no animation happens
+ * @typedef {Object} CameraOptions
+ * @property {Array} [center] Latitude and longitude (passed as `[lat, lng]`)
+ * @property {number} [zoom] Map zoom level
+ * @property {number} [bearing] Map rotation bearing in degrees counter-clockwise from north
+ * @property {number} [pitch] The angle at which the camera is looking at the ground
  */
-util.extend(exports, /** @lends Map.prototype */{
-    isEasing: function() {
-        return !!this._abortFn;
-    },
+
+/**
+ * @typedef {Object} [AnimationOptions]
+ * @property {number} [duration=500] Number in milliseconds
+ * @property {Function} [easing]
+ * @property {Array} [offset=[0,0]] point, origin of movement relative to map center
+ * @property {boolean} [animate=true] When set to false, no animation happens
+ */
+
+var Camera = module.exports = function() {};
+
+util.extend(Camera.prototype, /** @lends Map.prototype */{
+    /**
+     * Get the current view geographical point.
+     * @returns {LatLng}
+     */
+    getCenter: function() { return this.transform.center; },
 
     /**
-     * Stop current animation
+     * Sets a map location. Equivalent to `jumpTo({center: center})`.
      *
-     * @returns {this}
+     * @param {Array} center Latitude and longitude (passed as `[lat, lng]`)
+     * @fires movestart
+     * @fires moveend
+     * @returns {Map} `this`
+     * @example
+     * map.setCenter([-74, 38]);
      */
-    stop: function() {
-        if (this._abortFn) {
-            this._abortFn.call(this);
-            delete this._abortFn;
-
-            this._finishFn.call(this);
-            delete this._finishFn;
-        }
-        return this;
-    },
-
-    _ease: function(frame, finish, options) {
-        this._finishFn = finish;
-        this._abortFn = browser.timed(function (t) {
-            frame.call(this, options.easing(t));
-            if (t === 1) {
-                delete this._abortFn;
-                this._finishFn.call(this);
-                delete this._finishFn;
-            }
-        }, options.animate === false ? 0 : options.duration, this);
+    setCenter: function(center) {
+        this.jumpTo({center: center});
     },
 
     /**
      * Pan by a certain number of pixels
      *
      * @param {Array} offset [x, y]
-     * @param {animOptions}
+     * @param {AnimationOptions}
      * @fires movestart
      * @fires moveend
      * @returns {this}
@@ -65,7 +64,7 @@ util.extend(exports, /** @lends Map.prototype */{
      * Pan to a certain location with easing
      *
      * @param {Object} latlng a `LatLng` object
-     * @param {animOptions}
+     * @param {AnimationOptions}
      * @fires movestart
      * @fires moveend
      * @returns {this}
@@ -92,7 +91,7 @@ util.extend(exports, /** @lends Map.prototype */{
 
         this._ease(function(k) {
             tr.center = tr.unproject(from.add(to.sub(from).mult(k)));
-            this._move();
+            this.fire('move');
         }, function() {
             this.fire('moveend');
         }, options);
@@ -100,11 +99,33 @@ util.extend(exports, /** @lends Map.prototype */{
         return this;
     },
 
+
+    /**
+     * Get the current zoom
+     * @returns {number}
+     */
+    getZoom: function() { return this.transform.zoom; },
+
+    /**
+     * Sets a map zoom. Equivalent to `jumpTo({zoom: zoom})`.
+     *
+     * @param {number} zoom Map zoom level
+     * @fires movestart
+     * @fires moveend
+     * @returns {Map} `this`
+     * @example
+     * // zoom the map to 5
+     * map.setZoom(5);
+     */
+    setZoom: function(zoom) {
+        this.jumpTo({zoom: zoom});
+    },
+
     /**
      * Zooms to a certain zoom level with easing.
      *
      * @param {Number} zoom
-     * @param {animOptions}
+     * @param {AnimationOptions}
      * @fires movestart
      * @fires moveend
      * @returns {this}
@@ -137,8 +158,7 @@ util.extend(exports, /** @lends Map.prototype */{
 
         this._ease(function(k) {
             tr.setZoomAround(interpolate(startZoom, zoom, k), around);
-            this.animationLoop.set(300); // text fading
-            this._move(true);
+            this.fire('move').fire('zoom');
         }, function() {
             this.ease = null;
             if (options.duration >= 200) {
@@ -151,7 +171,6 @@ util.extend(exports, /** @lends Map.prototype */{
             clearTimeout(this._onZoomEnd);
             this._onZoomEnd = setTimeout(function() {
                 this.zooming = false;
-                this._rerender();
                 this.fire('moveend');
             }.bind(this), 200);
         }
@@ -162,7 +181,7 @@ util.extend(exports, /** @lends Map.prototype */{
     /**
      * Zoom in by 1 level
      *
-     * @param {animOptions}
+     * @param {AnimationOptions}
      * @fires movestart
      * @fires moveend
      * @returns {this}
@@ -174,7 +193,7 @@ util.extend(exports, /** @lends Map.prototype */{
     /**
      * Zoom out by 1 level
      *
-     * @param {animOptions}
+     * @param {AnimationOptions}
      * @fires movestart
      * @fires moveend
      * @returns {this}
@@ -183,11 +202,33 @@ util.extend(exports, /** @lends Map.prototype */{
         this.zoomTo(this.getZoom() - 1, options);
     },
 
+
+    /**
+     * Get the current bearing in degrees
+     * @returns {number}
+     */
+    getBearing: function() { return this.transform.bearing; },
+
+    /**
+     * Sets a map rotation. Equivalent to `jumpTo({bearing: bearing})`.
+     *
+     * @param {number} bearing Map rotation bearing in degrees counter-clockwise from north
+     * @fires movestart
+     * @fires moveend
+     * @returns {Map} `this`
+     * @example
+     * // rotate the map to 90 degrees
+     * map.setBearing(90);
+     */
+    setBearing: function(bearing) {
+        this.jumpTo({bearing: bearing});
+    },
+
     /**
      * Rotate bearing by a certain number of degrees with easing
      *
      * @param {Number} bearing
-     * @param {animOptions}
+     * @param {AnimationOptions}
      * @fires movestart
      * @fires moveend
      * @returns {this}
@@ -217,7 +258,7 @@ util.extend(exports, /** @lends Map.prototype */{
 
         this._ease(function(k) {
             tr.setBearingAround(interpolate(start, bearing, k), around);
-            this._move(false, true);
+            this.fire('move').fire('rotate');
         }, function() {
             this.rotating = false;
             this.fire('moveend');
@@ -229,7 +270,7 @@ util.extend(exports, /** @lends Map.prototype */{
     /**
      * Sets map bearing to 0 (north) with easing
      *
-     * @param {animOptions}
+     * @param {AnimationOptions}
      * @fires movestart
      * @fires moveend
      * @returns {this}
@@ -237,6 +278,26 @@ util.extend(exports, /** @lends Map.prototype */{
     resetNorth: function(options) {
         return this.rotateTo(0, util.extend({duration: 1000}, options));
     },
+
+
+    /**
+     * Get the current angle in degrees
+     * @returns {number}
+     */
+    getPitch: function() { return this.transform.pitch; },
+
+    /**
+     * Sets a map angle. Equivalent to `jumpTo({pitch: pitch})`.
+     *
+     * @param {number} pitch The angle at which the camera is looking at the ground
+     * @fires movestart
+     * @fires moveend
+     * @returns {Map} `this`
+     */
+    setPitch: function(pitch) {
+        this.jumpTo({pitch: pitch});
+    },
+
 
     /**
      * Zoom to contain certain geographical bounds
@@ -267,31 +328,82 @@ util.extend(exports, /** @lends Map.prototype */{
             nw = tr.project(bounds.getNorthWest()),
             se = tr.project(bounds.getSouthEast()),
             size = se.sub(nw),
-            center = tr.unproject(nw.add(se).div(2)),
-
             scaleX = (tr.width - options.padding * 2 - Math.abs(offset.x) * 2) / size.x,
-            scaleY = (tr.height - options.padding * 2 - Math.abs(offset.y) * 2) / size.y,
+            scaleY = (tr.height - options.padding * 2 - Math.abs(offset.y) * 2) / size.y;
 
-            zoom = Math.min(tr.scaleZoom(tr.scale * Math.min(scaleX, scaleY)), options.maxZoom);
+        options.center = tr.unproject(nw.add(se).div(2));
+        options.zoom = Math.min(tr.scaleZoom(tr.scale * Math.min(scaleX, scaleY)), options.maxZoom);
+        options.bearing = 0;
 
         return options.linear ?
-            this.easeTo(center, zoom, 0, this.getPitch(), options) :
-            this.flyTo(center, zoom, 0, options);
+            this.easeTo(options) :
+            this.flyTo(options);
+    },
+
+    /**
+     * Change any combination of center, zoom, bearing, and pitch, without
+     * a transition. The map will retain the current values for any options
+     * not included in `options`.
+     *
+     * @param {CameraOptions} options map view options
+     * @fires movestart
+     * @fires moveend
+     * @returns {Map} `this`
+     */
+    jumpTo: function(options) {
+        this.stop();
+
+        var tr = this.transform,
+            zoomChanged = false,
+            bearingChanged = false,
+            pitchChanged = false;
+
+        if ('center' in options) {
+            tr.center = LatLng.convert(options.center);
+        }
+
+        if ('zoom' in options && tr.zoom !== +options.zoom) {
+            zoomChanged = true;
+            tr.zoom = +options.zoom;
+        }
+
+        if ('bearing' in options && tr.bearing !== +options.bearing) {
+            bearingChanged = true;
+            tr.bearing = +options.bearing;
+        }
+
+        if ('pitch' in options && tr.pitch !== +options.pitch) {
+            pitchChanged = true;
+            tr.pitch = +options.pitch;
+        }
+
+        this.fire('movestart')
+            .fire('move');
+
+        if (zoomChanged) {
+            this.fire('zoom');
+        }
+
+        if (bearingChanged) {
+            this.fire('rotate');
+        }
+
+        if (pitchChanged) {
+            this.fire('pitch');
+        }
+
+        return this.fire('moveend');
     },
 
     /**
      * Easing animation to a specified location/zoom/bearing
      *
-     * @param {Object} latlng a `LatLng` object
-     * @param {Number} zoom
-     * @param {Number} bearing
-     * @param {Number} pitch
-     * @param {animOptions}
+     * @param {CameraOptions+AnimationOptions} options map view and animation options
      * @fires movestart
      * @fires moveend
      * @returns {this}
      */
-    easeTo: function(latlng, zoom, bearing, pitch, options) {
+    easeTo: function(options) {
         this.stop();
 
         options = util.extend({
@@ -302,25 +414,27 @@ util.extend(exports, /** @lends Map.prototype */{
 
         var tr = this.transform,
             offset = Point.convert(options.offset).rotate(-tr.angle),
+            from = tr.point,
             startZoom = this.getZoom(),
             startBearing = this.getBearing(),
             startPitch = this.getPitch();
 
-        latlng = LatLng.convert(latlng);
-        zoom = zoom === undefined ? startZoom : zoom;
-        bearing = bearing === undefined ? startBearing : this._normalizeBearing(bearing, startBearing);
-        pitch = pitch === undefined ? startPitch : pitch;
+        var zoom = 'zoom' in options ? +options.zoom : startZoom;
+        var bearing = 'bearing' in options ? this._normalizeBearing(options.bearing, startBearing) : startBearing;
+        var pitch = 'pitch' in options ? +options.pitch : startPitch;
 
         var scale = tr.zoomScale(zoom - startZoom),
-            from = tr.point,
-            to = latlng ? tr.project(latlng).sub(offset.div(scale)) : tr.point,
+            to = 'center' in options ? tr.project(LatLng.convert(options.center)).sub(offset.div(scale)) : from,
             around;
 
         if (zoom !== startZoom) {
             around = tr.pointLocation(tr.centerPoint.add(to.sub(from).div(1 - 1 / scale)));
             this.zooming = true;
         }
-        if (startBearing !== bearing) this.rotating = true;
+
+        if (startBearing !== bearing) {
+            this.rotating = true;
+        }
 
         this.fire('movestart');
 
@@ -339,8 +453,13 @@ util.extend(exports, /** @lends Map.prototype */{
                 tr.pitch = interpolate(startPitch, pitch, k);
             }
 
-            this.animationLoop.set(300); // text fading
-            this._move(zoom !== startZoom, bearing !== startBearing);
+            this.fire('move');
+            if (this.zooming) {
+                this.fire('zoom');
+            }
+            if (this.rotating) {
+                this.fire('rotate');
+            }
         }, function() {
             this.zooming = false;
             this.rotating = false;
@@ -353,21 +472,20 @@ util.extend(exports, /** @lends Map.prototype */{
     /**
      * Flying animation to a specified location/zoom/bearing with automatic curve
      *
-     * @param {Object} latlng a `LatLng` object
-     * @param {Number} zoom
-     * @param {Number} bearing
-     * @param {Object} options
+     * @param {CameraOptions} options map view options
      * @param {Number} [options.speed=1.2] How fast animation occurs
      * @param {Number} [options.curve=1.42] How much zooming out occurs during animation
-     * @param {Function} options.easing
+     * @param {Function} [options.easing]
      * @fires movestart
      * @fires moveend
      * @returns {this}
      * @example
      * // fly with default options to null island
-     * map.flyTo([0, 0], 9, 0);
+     * map.flyTo({center: [0, 0], zoom: 9});
      * // using flyTo options
-     * map.flyTo([0, 0], 9, 0, {
+     * map.flyTo({
+     *   center: [0, 0],
+     *   zoom: 9,
      *   speed: 0.2,
      *   curve: 1,
      *   easing: function(t) {
@@ -375,7 +493,7 @@ util.extend(exports, /** @lends Map.prototype */{
      *   }
      * });
      */
-    flyTo: function(latlng, zoom, bearing, options) {
+    flyTo: function(options) {
         this.stop();
 
         options = util.extend({
@@ -385,23 +503,18 @@ util.extend(exports, /** @lends Map.prototype */{
             easing: util.ease
         }, options);
 
-        latlng = LatLng.convert(latlng);
-
-        var offset = Point.convert(options.offset),
-            tr = this.transform,
+        var tr = this.transform,
+            offset = Point.convert(options.offset),
             startZoom = this.getZoom(),
             startBearing = this.getBearing();
 
-        zoom = zoom === undefined ? startZoom : zoom;
-        bearing = bearing === undefined ? startBearing : this._normalizeBearing(bearing, startBearing);
+        var center = 'center' in options ? LatLng.convert(options.center) : this.getCenter();
+        var zoom = 'zoom' in options ?  +options.zoom : startZoom;
+        var bearing = 'bearing' in options ? this._normalizeBearing(options.bearing, startBearing) : startBearing;
 
         var scale = tr.zoomScale(zoom - startZoom),
             from = tr.point,
-            to = tr.project(latlng).sub(offset.div(scale));
-
-        if (options.animate === false) {
-            return this.setView(latlng, zoom, bearing, this.getPitch());
-        }
+            to = tr.project(center).sub(offset.div(scale));
 
         var startWorldSize = tr.worldSize,
             rho = options.curve,
@@ -454,9 +567,10 @@ util.extend(exports, /** @lends Map.prototype */{
                 tr.bearing = interpolate(startBearing, bearing, k);
             }
 
-            this.animationLoop.set(300); // text fading
-
-            this._move(true, bearing !== startBearing);
+            this.fire('move').fire('zoom');
+            if (bearing !== startBearing) {
+                this.fire('rotate');
+            }
         }, function() {
             this.zooming = false;
             this.rotating = false;
@@ -464,6 +578,38 @@ util.extend(exports, /** @lends Map.prototype */{
         }, options);
 
         return this;
+    },
+
+    isEasing: function() {
+        return !!this._abortFn;
+    },
+
+    /**
+     * Stop current animation
+     *
+     * @returns {this}
+     */
+    stop: function() {
+        if (this._abortFn) {
+            this._abortFn.call(this);
+            delete this._abortFn;
+
+            this._finishFn.call(this);
+            delete this._finishFn;
+        }
+        return this;
+    },
+
+    _ease: function(frame, finish, options) {
+        this._finishFn = finish;
+        this._abortFn = browser.timed(function (t) {
+            frame.call(this, options.easing(t));
+            if (t === 1) {
+                delete this._abortFn;
+                this._finishFn.call(this);
+                delete this._finishFn;
+            }
+        }, options.animate === false ? 0 : options.duration, this);
     },
 
     // convert bearing so that it's numerically close to the current one so that it interpolates properly