+++ js/util/mru_cache.js
@@ -4,16 +4,22 @@
  * A [most-recently-used cache](http://en.wikipedia.org/wiki/Cache_algorithms)
  * with hash lookup made possible by keeping a list of keys in parallel to
  * an array of dictionary of values
+ *
+ * @param {number} max number of permitted values
+ * @param {Function} onRemove callback called with items when they expire
+ * @private
  */
 module.exports = MRUCache;
-function MRUCache(length, onRemove) {
-    this.max = length;
+function MRUCache(max, onRemove) {
+    this.max = max;
     this.onRemove = onRemove;
     this.reset();
 }
 
-/*
- * Clears the cache
+/**
+ * Clear the cache
+ *
+ * @returns {MRUCache} this cache
  */
 MRUCache.prototype.reset = function() {
     for (var key in this.list) {
@@ -29,6 +35,11 @@ MRUCache.prototype.reset = function() {
 /*
  * Add a key, value combination to the cache, trimming its size if this pushes
  * it over max length.
+ *
+ * @param {string} key lookup key for the item
+ * @param {*} data any value
+ *
+ * @returns {MRUCache} this cache
  */
 MRUCache.prototype.add = function(key, data) {
     this.list[key] = data;
@@ -44,6 +55,9 @@ MRUCache.prototype.add = function(key, data) {
 
 /*
  * Determine whether the value attached to `key` is present
+ *
+ * @param {string} key the key to be looked-up
+ * @returns {boolean} whether the cache has this value
  */
 MRUCache.prototype.has = function(key) {
     return key in this.list;
@@ -51,6 +65,8 @@ MRUCache.prototype.has = function(key) {
 
 /*
  * List all keys in the cache
+ *
+ * @returns {Array<string>} an array of keys in this cache.
  */
 MRUCache.prototype.keys = function() {
     return this.order;
@@ -59,6 +75,9 @@ MRUCache.prototype.keys = function() {
 /*
  * Get the value attached to a specific key. If the key is not found,
  * returns `null`
+ *
+ * @param {string} key the key to look up
+ * @returns {*} the data, or null if it isn't found
  */
 MRUCache.prototype.get = function(key) {
     if (!this.has(key)) { return null; }